Little rearrangement. Added dockerfile. Note that you will need to make the /climate/data/ folder a volume so that the soundings and other dbs persist past the container. I haven't checked the paths to the other programs to check for breakage. And the DB is not reading the .db file. I presume because the paths are not handled properly in the code. Which means that the code will need to be harden for this type of thing. I need to clean up the code anyway.
