+++ bin/study-cold_soundings/2016122500.70261.stuve10.parc.gif
