+++ bin/study-cold_soundings/Fig1-mbar_height_std_atm.PNG
