+++ bin/study-cold_soundings/cold_sounding.py
