+++ bin/study-cold_soundings/default.py
