+++ bin/study-cold_soundings/soundings_2016.png
