+++ bin/study-cold_soundings/soundings_temp_over_year_line.png
